pyproject.toml
[project]
name = "kalshi-bot"
version = "0.2.0"
description = "Automated Kalshi trading bot with LightGBM prediction model."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "httpx>=0.27.0",
  "cryptography>=43.0.0",
  "websockets>=12.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "PyYAML>=6.0.1",
  "structlog>=24.2.0",
  "numpy>=1.24.0",
  "lightgbm>=4.0.0",
  "streamlit>=1.36.0",
  "streamlit-autorefresh>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.8",
  "ruff>=0.5.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py39"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"
